Add ControlConnection.getHostsInfo() and .getClusterInfo()
ControlConnection heavily relies on static methods like refreshNodeListAndTokenMap().
It is easy to call, but makes flow complex, because it is difficult to find the order
of Host objects modifications and onUp()/onDown() invocations. Also this breaks
incapsulation, because any method of Cluster and its inner classes can be invoked
during refresh.

Instead we will refactor this into ControlConnection.getXXX() API and then we apply
gathered info to Cluster and Metadata. This commit introduces first getter methods.

Note: this commit has several TODO items regarding exception handling and some other
things and is not production-ready.